Initializing Repository
The project has been initialized with its basic configuration
generated by Spring Initializr. The pom.xml file already
has all the necessary dependencies used in this project
and the .gitignore file has been changed to ignore certain
files used to store sensitive settings such as .env and
application.properties. In addition, of course, to the folder that
stores the data eventually added to the database - api-volume.

IMPORTANT: Some of the files present in this project in the future will
only be useful if Docker containers are being used. However, if a real
database is used, they can be ignored - docker-compose.yaml.