build: Adds a docker-compose file for correct database container build
Added a properly configured docker-compose.yaml file to build a MySQL
database container. The .env file is also designed to hold certain
settings related to container environment variables and make them
hidden and secure when versioning to GitHub. Changes to the
application.properties file were necessary for the implementation of
the application with the database to occur correctly.