feat: Adds the Request and Response class to the Services API
The Request class is designed to show the user the model
and what data is required to have in the data object. In other
words, describes how the API should receive your data objects
for POST and PUT endpoints.

The Response class, on the other hand, was created to show what data is
returned by the API as a response when using the POST, PUT and, mainly,
GET endpoints. Shows a more complete view of the data object when
compared to Request.